Okay, it's all working the way I want...
Except for the fact that I can't get standalone to work. And unfortunate non-scripted changes are required to make blanket work. But it is a nice set of two files, and it works not just with my own Harness, but also with grunt-mocha's PhantomJS setup. Making that experience just a little bit nicer will round it out. Then I can write that task to pull things from diet into projects using this node module. Whew! What a week.